#80 - Reduced code duplication
